src/backend/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Cython==0.29.7
idna==2.8
marshmallow==2.19.2
meld3==1.0.2
multidict==4.5.2
mypy==0.701
mypy-extensions==0.4.1
PyYAML==5.1
raven==6.10.0
raven-aiohttp==0.7.0
sentry-sdk==0.7.14
supervisor==4.0.2
typed-ast==1.3.5
ujson==1.35
urllib3==1.25.2
uvloop==0.12.2
yarl==1.3.0
Wikipedia-API==0.5.1
spacy==2.1.4
msgpack==0.6.1